requirements.txt
pandas
polars
python-dotenv
facebook-business
google-analytics-data
google-ads-api-report-fetcher